ATSbackend/requirements.txt
# Core API
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
slowapi>=0.1.9

# Document parsing
pdfplumber==0.11.0
python-docx==1.1.2

# SLM skill extraction
# On first run downloads Qwen2-0.5B (~400 MB) to ~/.cache/huggingface
# Set FORCE_REGEX=1 to skip; set MIN_FREE_MB=0 to always attempt
transformers>=4.40.0
torch>=2.2.0
accelerate>=0.29.0
tf-keras

# spaCy Tier-2 fallback (~15 MB)
spacy>=3.7.0

# Embeddings + ML pipeline
sentence-transformers==3.0.1
scikit-learn==1.5.0
joblib>=1.3.0
xgboost>=2.0.0
networkx==3.3
numpy==1.26.4
scipy==1.13.1

# Soft/fuzzy matching
rapidfuzz>=3.0.0
